consolidate some packages
