+++ server/server.go
@@ -21,14 +21,14 @@ import (
 	"github.com/hacdias/eagle/core"
 	"github.com/hacdias/eagle/hooks"
 	"github.com/hacdias/eagle/log"
-	"github.com/hacdias/eagle/media"
 	"github.com/hacdias/eagle/pkg/maze"
 	"github.com/hacdias/eagle/services/bunny"
 	"github.com/hacdias/eagle/services/imgproxy"
+	"github.com/hacdias/eagle/services/media"
 	"github.com/hacdias/eagle/services/miniflux"
 	"github.com/hacdias/eagle/services/postgres"
 	"github.com/hacdias/eagle/services/telegram"
-	"github.com/hacdias/eagle/webmentions"
+	"github.com/hacdias/eagle/services/webmentions"
 	"github.com/hacdias/indieauth/v3"
 	"github.com/hashicorp/go-multierror"
 	"github.com/robfig/cron/v3"
